requirements.txt
cachetools==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
coreapi==2.3.1
coreschema==0.0.4
defusedxml==0.5.0
dill==0.2.7.1
Django==1.10.7
django-allauth==0.32.0
django-endless-pagination==2.0
django-haystack==2.6.1
django-registration==2.2
django-registration-email==0.9
django-rest-auth==0.9.1
django-rest-swagger==2.0.7
django-social-auth==0.7.28
djangorestframework==3.5.1
Flask==0.12.2
Flask-PyMongo==0.5.1
Flask-SQLAlchemy==2.2
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
gapic-google-cloud-pubsub-v1==0.15.4
gapic-google-cloud-spanner-admin-database-v1==0.15.3
gapic-google-cloud-spanner-admin-instance-v1==0.15.3
gapic-google-cloud-spanner-v1==0.15.3
google-auth==1.0.2
google-cloud==0.27.0
google-cloud-bigquery==0.26.0
google-cloud-bigtable==0.26.0
google-cloud-core==0.26.0
google-cloud-datastore==1.2.0
google-cloud-dns==0.26.0
google-cloud-error-reporting==0.26.0
google-cloud-language==0.27.0
google-cloud-logging==1.2.0
google-cloud-monitoring==0.26.0
google-cloud-pubsub==0.27.0
google-cloud-resource-manager==0.26.0
google-cloud-runtimeconfig==0.26.0
google-cloud-spanner==0.26.0
google-cloud-speech==0.28.0
google-cloud-storage==1.3.2
google-cloud-translate==1.1.0
google-cloud-videointelligence==0.25.0
google-cloud-vision==0.26.0
google-gax==0.15.14
google-resumable-media==0.2.3
googleapis-common-protos==1.5.2
grpc-google-iam-v1==0.11.1
grpcio==1.4.0
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
monotonic==1.3
mysqlclient==1.3.10
numpy==1.14.0
oauth2==1.9.0.post1
oauth2client==4.1.2
oauthlib==2.0.2
olefile==0.44
openapi-codec==1.3.2
Pillow==4.2.1
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
proto-google-cloud-pubsub-v1==0.15.4
proto-google-cloud-spanner-admin-database-v1==0.15.3
proto-google-cloud-spanner-admin-instance-v1==0.15.3
proto-google-cloud-spanner-v1==0.15.3
protobuf==3.4.0
pyasn1==0.3.2
pyasn1-modules==0.0.11
pymongo==3.5.0
PyMySQL==0.7.11
python-openid==2.2.5
python3-openid==3.1.0
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
simplejson==3.11.1
six==1.9.0
SQLAlchemy==1.1.13
tenacity==4.4.0
uritemplate==3.0.0
urllib3==1.22
Werkzeug==0.12.2
